+++ packages/quantic/force-app/main/default/lwc/quanticSearchBox/__tests__/quanticSearchBox.test.js
@@ -1,0 +1,15 @@
+import {createElement} from 'lwc';
+import QuanticSearchBox from 'c/quanticSearchBox';
+import {cleanup} from 'c/testUtils';
+
+describe('c-quantic-search-box', () => {
+  afterEach(() => {
+    cleanup();
+  });
+
+  it('construct itself without throwing', () => {
+    expect(() =>
+      createElement('c-quantic-search-box', {is: QuanticSearchBox})
+    ).not.toThrow();
+  });
+});