+++ packages/quantic/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.10.0-alpha.4](https://bitbucket.org/coveord/ui-kit/compare/v0.10.0-alpha.3...v0.10.0-alpha.4) (2021-03-31)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+# [0.10.0-alpha.3](https://bitbucket.org/coveord/ui-kit/compare/v0.10.0-alpha.2...v0.10.0-alpha.3) (2021-03-31)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+# [0.10.0-alpha.2](https://bitbucket.org/coveord/ui-kit/compare/v0.10.0-alpha.1...v0.10.0-alpha.2) (2021-03-31)
+
+
+### Features
+
+* **quantic:** renamed quantic components to avoid naming conflicts ([#657](https://bitbucket.org/coveord/ui-kit/issues/657)) ([2ea04ea](https://bitbucket.org/coveord/ui-kit/commits/2ea04eaf04ea8a2cbc550231b15084f0265684c0))
+
+
+
+
+
+# [0.10.0-alpha.0](https://bitbucket.org/coveord/ui-kit/compare/v0.9.0...v0.10.0-alpha.0) (2021-03-30)
+
+
+### Features
+
+* **quantic:** removed atomic dependency ([#652](https://bitbucket.org/coveord/ui-kit/issues/652)) ([0814a26](https://bitbucket.org/coveord/ui-kit/commits/0814a267c0236e10d5c545a77850943b14094f98))
+
+
+
+
+
 # [0.9.0](https://bitbucket.org/coveord/ui-kit/compare/v0.9.0-alpha.9...v0.9.0) (2021-03-26)
 
 **Note:** Version bump only for package @coveo/quantic